refactor: move PackGenerator to btrace-core to break instr->compiler compile dependency
PackGenerator interface moved from io.btrace.compiler to io.btrace.core so that
btrace-instr no longer has a compile-time dependency on btrace-compiler; the compiler
is retained as a runtimeOnly dep in btrace-instr solely for the compileTestProbes task.
