+++ btrace-core/src/main/java/io/btrace/core/PackGenerator.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.btrace.compiler;
+package io.btrace.core;
 
 import java.io.IOException;
