+++ btrace-instr/src/main/resources/META-INF/services/io.btrace.core.PackGenerator
